requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
csvw==3.5.1
cycler==0.12.1
dataclasses-json==0.6.7
diskcache==5.6.3
dlinfo==2.0.0
emoji==2.14.1
espeakng-loader==0.2.4
exceptiongroup==1.3.0
facenet-pytorch==2.6.0
faiss-cpu==1.11.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
isodate==0.7.2
jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kokoro-onnx==0.4.5
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
language-tags==1.2.0
llama-cpp-python==0.3.9
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mediapipe==0.10.7
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.3
onnxruntime==1.21.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
orjson==3.10.18
packaging==24.2
phonemizer-fork==3.3.1
pillow==11.2.1
propcache==0.3.1
protobuf==3.20.3
pyaudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
segments==2.3.0
sentence-transformers==4.1.0
setuptools==57.4.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
sqlalchemy==2.0.41
srt==3.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.4.0
vosk==0.3.45
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0


# --- Dépendances spécifiques à Linux (et nécessitant compilation/librairies système) ---
Adafruit-Blinka==8.58.1; sys_platform == "linux"
adafruit-circuitpython-busdevice==5.2.12; sys_platform == "linux"
adafruit-circuitpython-connectionmanager==3.1.4; sys_platform == "linux"
adafruit-circuitpython-requests==4.1.10; sys_platform == "linux"
adafruit-circuitpython-typing==1.11.2; sys_platform == "linux"
Adafruit-PlatformDetect==3.79.0; sys_platform == "linux"
Adafruit-PureIO==1.1.11; sys_platform == "linux"
binho-host-adapter==0.1.6; sys_platform == "linux"
gpiod==2.3.0; sys_platform == "linux"
pyaudio==0.2.14; sys_platform == "linux"
pyftdi==0.56.0; sys_platform == "linux"
pyserial==3.5; sys_platform == "linux"
pyusb==1.3.1; sys_platform == "linux"
sysv_ipc==1.1.0; sys_platform == "linux"


#requirement supplémentaire
audioread==3.0.1
decorator==5.2.1
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
msgpack==1.1.0
numba==0.61.2
platformdirs==4.3.8
pooch==1.8.2
soxr==0.5.0.post1

